requirements.txt
# AI Safety Surveillance — PPE Compliance Monitor
# Python Dependencies

# Core ML / Vision
ultralytics>=8.2.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
numpy>=1.24.0

# Web Framework
Flask>=3.0.0
flask-socketio>=5.3.0

# SocketIO async support
eventlet>=0.35.0

# Utilities
PyYAML>=6.0
Pillow>=10.0.0

# Dataset download (optional)
# roboflow>=1.1.0

# Production server
gunicorn>=21.0.0
